[ADD] point_of_sale: added salesperson field
Introduced the functionality to assign a salesperson to a Point of Sale order.

Added new dialog menu, populated with employees from hr.employee, has been added
 to the main POS interface.

A  (Many2one) field was added to the  model to store
 the selected employee.

The POS Order form and list views in the backend have been updated to display
the salesperson's name.

Employee data is now loaded into the POS session to make it available in the UI.

task-4968717