[ADD] sales_purchase_barcode: implement barcode scanning in product catalog
Before this commit:

Users had to manually enter the barcode into the search bar to find products
in the catalog view.

After this commit:

Scanning a barcode automatically adds a quantity of 1 to the order without
manual intervention.

Displays a toaster message saying “No product found with this barcode number” if
no matching product is found.

If the same product is scanned multiple times, its quantity is incremented on
the existing order line.

All scanned products are sorted and displayed on the first page for easy access.