[IMP] Estate_property: computed and onchange functionality
provided the the automatic computed value for the total_area field by suming garden_area and living_area.
Also computed the value for the best_price based on the maximum price in the offers.
Provided the onchange functionality on garden. if the value is set than default value for garden_area and garden_orientation are set.

[IMP] Estate_property_offer: computed the validity and deadline_date

based on the values of the deadline or validity the value changes acoordingly.