[ADD] estate: Added the constraints and accordingly provided validation message
Added the sql constraints for the for selling_price expected_price and offer_price
for all this the value must be positive. For the property type and tags the value
must be unique. Also added the constraints so that the selling price cannot be
lower than 90% of the expected price.