[ADD] estate: Added inline views, widgets, list order, attributes and buttons
Added the inline views so that the specific list of properties should be added to
the property type view.
Added the widget for the status bar for the stage of the property
Added the ordering of the properties, type, tag and offer
Added attributes and button where tags color can be changed it cannot be create and
edit from property page. conditionally display of sold and cancel button.
Added stat_button so that the offercount get visible on property type page and can
move to the offers by clicking on it