[ADD] estate: Added model and view inheritance on the estate model
Prevent deletion of a property if its state is not New or Cancelled.
set the property state to Offer Received when offer created and also
user cannot create an offer with a lower amount than an existing offer.
Added the the properties section for the user profile which display
all the property for that user. Also provided the view of properties
on user form.