[IMP] awesome_owl: make Card component generic with slot and add toggle feature
The Card component is improved to allow any kind of content inside it, not just
text. This means we can now put other components, like a counter, inside the
card easily.

Additionally, a button was added to the card header to show or hide the card
content. This makes the card more interactive and flexible to use.

Prop validation was also introduced to make sure the component receives the
right data.

Overall, these changes make the Card component more versatile and user-friendly.