[REF] estate: change property state to be a computed field.
Until now the state of a property is handled with manual assignments when it could be easily computed by in which state the property is in. This reduces complexity when adding new features on the matter of having to handle the state manually.
The change includes the creation of two variables (which should not be used outside the estate property file) to handle if a property is sold or cancelled.