[ADD] pos_workflow:  Allow creating backend orders with deferred payment
This commit introduces a new module pos_workflow which enhances the PoS flow by adding a new button "Place Order" in the PoS frontend.
A new action is available on the Product Screen to create a backend pos.order directly in pay later state along with the corresponding stock picking.
Orders created via this button are marked as pay later, allowing the user to defer payment.
Such orders remain visible in the PoS order list so the user can later select them and complete payment.