[ADD] purchase_discount: add global discount functionality in purchase order
Purpose:
When you have a big RFQ (with a lot of lines) and you receive a Quote from the
vendor with a global discount (% or amount), it's very time consuming to apply
it to all lines.

Technical changes:
Add discount_type, discount_in_value and computed discount_in_percentage fields
on purchase.order.
Add constraints for validate discount percentage.
Add discount button in purchase order view form.
Add action_discount to open a popup and apply_discount to apply the global
discount on all order lines.
Discount can be entered either as a percentage or as a fixed amount.

task-5366796