[IMP] estate:  enable UI interaction trough menus, display property data in list view and create/edit property data in form view
The model was created and access rights provided but there was no interaction with the model.  This is implemented through xml as its is mre convinient for complex data. There is one menu to acces the list view of the properties and a form views is displayed whend creating or editing a property. Some fields have now a default value, some are read-only, some are requiered and some cannot be copied when duplicated. The fields active and state were added for properties display in the list view by default and for it to have a different state depending on the stage respectively.