[IMP] estate: enrich the models with new fields (best price and total area in estate property and validity and deadline in property offer).
Best price is computed on the maximum price of all the offers linked to the property.

Total area is computed as the sum of the lving area and garden area.

Validity is the number days that have passed since the creation of the offer during wich it remains valid.

Deadline is the date until wich the offer remains valid. It is computed based on validity and when changed, updates validity.