[IMP] estate: allow more direct interactions with the property and its offers.
The state of a property and its offer could only be changed manually which is not safe way of handling those status.

Now buttons have been added : sold and cancel for the property and accept and refuse for the offers. An error pop-up is shown when trying to perfom an illegal action.