[IMP] awesome_dashboard: add a layout and reusable items to the dashboard.
This will add a new layout to the dashboard, new buttons to allow quick navigation to the customers kanban view and the leads, and create a new DashboardItem component.
This component is reusable and is displayed as a card on the dashboard. Its content is defined using a default slot to allow it to be set in the parent component.
The width of the card is defined as (18*size)rem, where size is a property of the class with a default value of 1. This allow to easily create dashboard items of different sizes.