[IMP] awesome_dashboard: add realtime update to dashboard data and dashboard lazy loading.
This will update the service so that it will update in realtime using a defined interval of 10 mintues (set as 10 seconds for test purpose).
To update the data displayed in the dashboard, the service return a reactive object.
This will also change the dashboard component from  an action to a lazy component. This will allow the user to only load the dashboard when it is displayed.
All dashboard item were moved into a new dashboard folder to make it easier to bundle and a new DashboadLoader component has been created to load the dashboard as a lazy component.