[IMP] awesome_dashboard: make the dashboard generic with item loaded from registry.
This will make the dashboard generic by using a loop on an items list into the xml template.
The item are fetch from a registry to allow other module to create dashboard item. This module load its items from the dashboard_items.js file.
This will also add a filtering dialog, that will allow the user to choose which dashboad item to diplay. The configuration is store locally into the browser storage.