[IMP] estate: imporove user interface with some business logic and buttons.
State are now displayed in a status bar. The buttons (Sold/cancel on properties and Accept/Refuse on offers) are now dynamic and only appeares when needed. Offers can be created and edited in place. Tags have colors. Property type can only be created and edited from the settings. Offers related to a  property type can be acces from its form.