[IMP] last_purchased_products: added customer name next to invoice time
Before, there was no customer name next to last invoice time,
it was hard to identify that to which customer this product was
recently invoiced.

this commit add that functionality which will add customer name
to whom then recent invoice  created and that name is clickable
will take to that invoice form.

Technical Details:
- Model `product.product`: Created `invoice_partner_name` Char field
  to store customer name.
- Model `product.product`: Created action `action_open_last_invoice`
  to open last invoice of product.
- View `product_views`: Updated to show customer name next to
  `last_invoice_time`.