[IMP] estate: alter the standart delete method of properties, the standart creation method of offers and the view form of an user.
The new delete method with the decorator 'ondelete()' prevent the deletion of a property that is not in the state 'New' or 'Cancelled'.

The new creation method prevent user from creatin an offer with a lower amount than an existing offer and change the state of the property to 'Offer recieved' when it was 'New'.

The list of available properties is now displayed in a new page on the user form view.